picross: the provenance index is a uint16_t, because 531 imports made 532 rows
The generator's own guard caught this on the first real corpus, which is what
it was for. `Puzzle::provenance` was a uint8_t on the reasoning that origins
dedup -- and they do, for artwork this fork draws: 68 puzzles, one row. They do
not for a collection. Every imported puzzle carries its OWN source URL, and the
URL is the field that actually identifies which puzzle to ask permission about,
so a 531-puzzle import produced 532 distinct rows and the generator refused to
emit rather than wrap the index into a table it does not fit.

Four bytes per puzzle on the device (36 -> 40, the padding rather than the
field), 272 bytes across the shipping bank. Correctness over that, easily: the
failure it buys off is a puzzle whose recorded author is somebody else's.

Verified on the full corpus locally: pictures.txt plus the 531 imported janko
puzzles generate a 599-puzzle bank -- 22 at 5x5, 291 at 10x10, 286 at 15x15,
three contiguous size runs, 532 provenance rows -- and host-tests/picross and
host-tests/ui are green over it. The worst clue load in that bank (8 runs in a
row AND a column, against 5 and 3 in the fork's own) renders on the simulator
with no clipping and no elided clue: the gutters size themselves from
kMaxRowRuns/kMaxColRuns and the 15x15 cell drops to 19px, which still fits.

None of that data is committed and none of it is in this tree. It lives in the
session scratchpad and is deleted with it.

Docs: none affected -- the field's width is stated where it is declared and
the header is generated. Triage: no card. Dream check: unchanged.

+++ src/apps_local/picross/PicrossPuzzles.h
@@ -48,7 +48,7 @@ constexpr int kProvenanceCount = static_cast<int>(sizeof(kProvenances) / sizeof(
 struct Puzzle {
   const char* name;
   uint8_t size;             // 5, 10 or 15
-  uint8_t provenance;       // index into kProvenances
+  uint16_t provenance;      // index into kProvenances
   uint16_t rows[kMaxSize];  // bit c set when (row, col) is solid
 };
